add mpd backend support
