+++ gomu.go
@@ -20,7 +20,7 @@ type Gomu struct {
 	playingBar *PlayingBar
 	queue      *Queue
 	playlist   *Playlist
-	player     *player.Player
+	player     player.Player
 	pages      *tview.Pages
 	colors     *Colors
 	command    Command
@@ -49,11 +49,19 @@ func newGomu() *Gomu {
 // constructor function `newGomu` so that we can
 // test independently
 func (g *Gomu) initPanels(app *tview.Application, args Args) {
+	var err error
 	g.app = app
 	g.playingBar = newPlayingBar()
 	g.queue = newQueue()
 	g.playlist = newPlaylist(args)
-	g.player = player.New(g.anko.GetInt("General.volume"))
+	// g.player = player.New(g.anko.GetInt("General.volume"))
+	// g.player, err = player.NewPlayer(g.anko.GetInt("General.volume"), "mpd")
+	g.player, err = player.NewPlayer(g.anko.GetInt("General.volume"), g.anko.GetString("General.backend_server"), g.anko.GetString("General.mpd_port"))
+	// g.player, err = player.NewPlayer(g.anko.GetInt("General.volume"), "beep")
+	if err != nil {
+		errorPopup(err)
+	}
+
 	g.pages = tview.NewPages()
 	g.panels = []Panel{g.playlist, g.queue, g.playingBar}
 }
@@ -144,6 +152,11 @@ func (g *Gomu) quit(args Args) error {
 		if err != nil {
 			return tracerr.Wrap(err)
 		}
+	}
+
+	err := gomu.player.Stop()
+	if err != nil {
+		return tracerr.Wrap(err)
 	}
 
 	gomu.app.Stop()
